--- WssElixir/app/src/main/java/com/example/wsselixir/HomeActivity.kt
@@ -1,11 +1,0 @@
-package com.example.wsselixir
-
-import android.os.Bundle
-import androidx.appcompat.app.AppCompatActivity
-
-class HomeActivity : AppCompatActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_home)
-    }
-}
