+++ WssElixir/app/src/main/java/com/example/wsselixir/presentation/home/HomeActivity.kt
@@ -1,0 +1,16 @@
+package com.example.wsselixir.presentation.home
+
+import android.os.Bundle
+import androidx.appcompat.app.AppCompatActivity
+import com.example.wsselixir.R
+import com.example.wsselixir.databinding.ActivityHomeBinding
+
+class HomeActivity : AppCompatActivity() {
+    private lateinit var binding: ActivityHomeBinding
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        binding = ActivityHomeBinding.inflate(layoutInflater)
+        setContentView(binding.root)
+    }
+}
