+++ WssElixir/app/src/main/java/com/example/wsselixir/presentation/myinformation/MyInformationActivity.kt
@@ -1,7 +1,8 @@
-package com.example.wsselixir
+package com.example.wsselixir.presentation.myinformation
 
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
+import com.example.wsselixir.R
 
 class MyInformationActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
